Switch release step back to standard agents

+++ .buildkite/steps/github-release.sh
@@ -12,6 +12,12 @@ then
 	exit 1
 fi
 
+if [ -z "${GITHUB_RELEASE_ACCESS_TOKEN}" ]
+then
+  echo "Error: Missing \$GITHUB_RELEASE_ACCESS_TOKEN"
+  exit 1
+fi
+
 echo '--- Downloading releases'
 
 rm -rf pkg
@@ -20,4 +26,6 @@ buildkite-agent artifact download "pkg/*" .
 
 echo '--- Creating GitHub Release'
 
-github-release "${VERSION}" pkg/* --commit "$(git rev-parse HEAD)" --github-repository "buildkite/elastic-ci-stack-s3-secrets-hooks"
+export GITHUB_RELEASE_REPOSITORY="buildkite/elastic-ci-stack-s3-secrets-hooks"
+
+github-release "${VERSION}" pkg/* --commit "$(git rev-parse HEAD)"
